Move the schema factories ./src/forms -> ./src/forms/factory
